build: add "numpy" to build dependencies
Not sure that's the right way to do it, but without it, cmake complains
that it can't find numpy and doesn't build ggcmpy, unless one builds
it editable with no build isolation.

This way, everything works, I'm just confused why I have to add numpy to
build dependencies when it's already listed as a dependency.