jrrle: keep track of opening file multiple times
Ideally, this shouldn't happen, and really the CachingFileManager should
prevent it -- but it doesn't. Fortran doesn't allow for the same
file being opened twice, so now we keep track of an open count and only
actual close it on the Fortran end once that's back down to zero.

However, since there's a single underlying Fortran file, there is the
chance that something that relies on state (like the current position in
the file) may break, as that state exists only once per file, no matter
how many time it's been opened.