f2py: interface epoch1966() and cotr()
epoch1966 not really all that useful, since this is more or less trivially
done with built-in python datetime{,64}, but I guess it's somewhat useful
as validation of the Fortran implementation.

cotr, however, should be quite useful to be able to use the same coordinate
transforms used in the OpenGGCM in post-processing.